Build the isBalanced method
I've had a hard time with this one, because I didn't want to modify one
of the ForEach methods I've built becaus they only call the clalback
function they are given but do not retrieve in any way the results. So
in the end I've changed the inOrderForEach method so that it returns an
array with the results of the callback (here the isBalancedFunc).
Also I encountered the problem of the this definition : I was trying to
use the height method inside the callback function. But I've found that
the "this" changes when it is called by a callback. So an easy way to
fix that is to create a const self = this, before invoking the callback
and changing all the this to self in the callback. Pfiouuu